fix: move pokemon list to its own screen
