+++ app/app/index.tsx
@@ -1,53 +1,15 @@
-import { sortObjectsByKey } from "@/utils";
-import { useEffect, useState } from "react";
-import { View, Text, StyleSheet } from "react-native";
-import usePokemons from "./hooks/usePokemons";
-import { PokemonDetail } from "./domain/pokemon.model";
-import PokemonTile from "./components/PokemonTile";
+import { useReactQueryDevTools } from "@dev-plugins/react-query";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import PokeList from "./screens/PokeList";
+
+const queryClient = new QueryClient();
 
 export default function HomeScreen() {
-  const [favorites, setFavorites] = useState<Array<PokemonDetail>>([]);
-  const [nonFavorites, setNonFavorites] = useState<Array<PokemonDetail>>([]);
-
-  const { data: pokemons, error } = usePokemons();
-
-  useEffect(() => {
-    if (pokemons) {
-      const organizedPokemons = sortObjectsByKey(pokemons, "height");
-      setNonFavorites(organizedPokemons);
-    }
-  }, [pokemons]);
-
-  const toggleFavorite = (pokemon: PokemonDetail) => {
-    if (favorites.find((p) => p.id === pokemon.id)) {
-      setFavorites(favorites.filter((p) => p.id !== pokemon.id));
-      setNonFavorites(sortObjectsByKey([...nonFavorites, pokemon], "height"));
-    } else {
-      setNonFavorites(nonFavorites.filter((p) => p.id !== pokemon.id));
-      setFavorites(sortObjectsByKey([...favorites, pokemon], "height"));
-    }
-  };
+  useReactQueryDevTools(queryClient);
 
   return (
-    <View style={styles.container}>
-      <Text style={styles.center}>No favoritos</Text>
-      {favorites.map((pokemon) => (
-        <PokemonTile
-          key={pokemon.id}
-          pokemon={pokemon}
-          onPress={toggleFavorite}
-        />
-      ))}
-    </View>
+    <QueryClientProvider client={queryClient}>
+      <PokeList />
+    </QueryClientProvider>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    justifyContent: "center",
-    flex: 1,
-  },
-  center: {
-    textAlign: "center",
-  },
-});
